+++ R/arc-select.R
@@ -272,6 +272,11 @@ obj_check_layer <- function(x,
   )
 }
 
+#' @noRd
+obj_is_layer <- function(x) {
+  rlang::inherits_any(x, c("FeatureLayer", "Table"))
+}
+
 #' Check if an object inherits from a set of classes
 #'
 #' [check_inherits_any()] wraps [rlang::inherits_any()] to error if an object